requirements.txt
beautifulsoup4 == 4.8.2
html5lib == 1.0.1
feedgenerator == 1.9.1
requests == 2.22.0
python-dateutil == 2.7.3
